Move 'http' and 'ftp' packages under 'protocol' package.
