+++ setup.py
@@ -28,16 +28,16 @@ StrictVersion(version)
 
 PROJECT_PACKAGES = [
     'wpull',
-    'wpull.abstract',
+    'wpull.protocol.abstract',
     'wpull.backport',
     'wpull.coprocessor',
     'wpull.database',
     'wpull.document',
     'wpull.document.htmlparse',
     'wpull.driver',
-    'wpull.ftp',
-    'wpull.ftp.ls',
-    'wpull.http',
+    'wpull.protocol.ftp',
+    'wpull.protocol.ftp.ls',
+    'wpull.protocol.http',
     'wpull.processor',
     'wpull.proxy',
     'wpull.recorder.',
