+++ wpull/document/html_test.py
@@ -6,7 +6,7 @@ from lxml.etree import _Element as Element
 from wpull.document.base_test import CODEC_NAMES, EBCDIC
 from wpull.document.html import HTMLReader
 from wpull.document.htmlparse.lxml_ import HTMLParser
-from wpull.http.request import Request, Response
+from wpull.protocol.http.request import Request, Response
 from wpull.url import URLInfo
