+++ wpull/processor/ftp.py
@@ -16,8 +16,8 @@ from wpull.backport.logging import BraceMessage as __
 from wpull.body import Body
 from wpull.cache import LRUCache
 from wpull.errors import ProtocolError
-from wpull.ftp.request import Request, ListingResponse
-from wpull.ftp.util import FTPServerError
+from wpull.protocol.ftp.request import Request, ListingResponse
+from wpull.protocol.ftp.util import FTPServerError
 from wpull.hook import Actions
 from wpull.item import LinkType
 from wpull.processor.base import BaseProcessor, BaseProcessorSession, \
