+++ wpull/protocol/abstract/stream.py
