+++ wpull/protocol/http/client_test.py
@@ -5,12 +5,12 @@ import io
 import warnings
 
 from trollius import From
-from wpull.abstract.client import DurationTimeout
+from wpull.protocol.abstract.client import DurationTimeout
 
 from wpull.connection import ConnectionPool, Connection
 from wpull.errors import NetworkError
-from wpull.http.client import Client
-from wpull.http.request import Request
+from wpull.protocol.http.client import Client
+from wpull.protocol.http.request import Request
 from wpull.recorder.base import BaseRecorder, BaseRecorderSession
 from wpull.testing.badapp import BadAppTestCase
 import wpull.testing.async
