+++ wpull/protocol/http/redirect.py
