+++ wpull/recorder/warc_test.py
@@ -6,12 +6,12 @@ import re
 
 from wpull.body import Body
 from wpull.database.sqltable import URLTable
-from wpull.http.request import Request as HTTPRequest, Response as HTTPResponse
+from wpull.protocol.http.request import Request as HTTPRequest, Response as HTTPResponse
 from wpull.recorder.base_test import BaseRecorderTest
 from wpull.recorder.warc import WARCRecorder, WARCRecorderParams
 from wpull.warc import WARCRecord
 import wpull.util
-from wpull.ftp.request import Request as FTPRequest, Response as FTPResponse, Reply as FTPReply
+from wpull.protocol.ftp.request import Request as FTPRequest, Response as FTPResponse, Reply as FTPReply
 
 
 _logger = logging.getLogger(__name__)
